+++ ucom_decoder/ucom_message.hpp
@@ -34,6 +34,6 @@ class UcomMessage : public json {
     int get_id() { return _message_id; }
 
     std::string get_header() const { return _header; }
-    int get_signal_count() const { return _signals.size(); } 
+    size_t get_signal_count() const { return _signals.size(); } 
     const std::vector<ucom_signal_ptr_t> &get_signals();
 };
